+++ flutter/lib/src/sentry_flutter.dart
@@ -15,6 +15,7 @@ import 'file_system_transport.dart';
 import 'flutter_exception_type_identifier.dart';
 import 'frame_callback_handler.dart';
 import 'integrations/connectivity/connectivity_integration.dart';
+import 'integrations/frames_tracking_integration.dart';
 import 'integrations/integrations.dart';
 import 'integrations/native_app_start_handler.dart';
 import 'integrations/screenshot_integration.dart';
@@ -23,7 +24,6 @@ import 'native/native_scope_observer.dart';
 import 'native/sentry_native_binding.dart';
 import 'profiling.dart';
 import 'renderer/renderer.dart';
-import 'span_frame_metrics_collector.dart';
 import 'version.dart';
 import 'view_hierarchy/view_hierarchy_integration.dart';
 
@@ -133,13 +133,6 @@ mixin SentryFlutter {
 
     options.addEventProcessor(PlatformExceptionEventProcessor());
 
-    // Disabled for web, linux and windows until we can reliably get the display refresh rate
-    if (options.platformChecker.platform.isAndroid ||
-        options.platformChecker.platform.isIOS ||
-        options.platformChecker.platform.isMacOS) {
-      options.addPerformanceCollector(SpanFrameMetricsCollector(options));
-    }
-
     _setSdk(options);
   }
 
@@ -154,6 +147,11 @@ mixin SentryFlutter {
 
     // Will call WidgetsFlutterBinding.ensureInitialized() before all other integrations.
     integrations.add(WidgetsFlutterBindingIntegration());
+
+    options.logger(
+      SentryLevel.info,
+      '`WidgetsBindingIntegration` is available in multi-view applications.',
+    );
 
     // Use PlatformDispatcher.onError instead of zones.
     if (isOnErrorSupported) {
@@ -176,6 +174,7 @@ mixin SentryFlutter {
         integrations.add(LoadContextsIntegration(native));
       }
       integrations.add(LoadImageListIntegration(native));
+      integrations.add(FramesTrackingIntegration(native));
       options.enableDartSymbolication = false;
     }